add categories adding, deleting and updating

+++ src/service/data-services/category.js
@@ -51,6 +51,26 @@ class CategoryService {
   }
 
   /**
+   * Проверяем уникальный ли заголовок для категории.
+   * @async
+   * @param {String} title - заголовок категории
+   * @return {Boolean} - вернет true, если категория уникальна, если нет - ошибку
+   */
+  async isTitleUnique(title) {
+    const category = await this._categoryModel.findOne({
+      where: {
+        title
+      }
+    });
+    console.log(category);
+    if (category) {
+      throw new Error(`Category is not unique`);
+    }
+    return true;
+  }
+
+
+  /**
    * Добавление комментария к статье
    * @async
    * @param {{title: String}} categoryData - статья
@@ -67,6 +87,30 @@ class CategoryService {
    * @param {Number} categoryId - id категории
    */
   async delete(categoryId) {
+    const category = await this._categoryModel.findByPk(categoryId, {
+      attributes: [
+        `id`,
+        [
+          Sequelize.fn(
+              `COUNT`,
+              `*`
+          ),
+          `count`
+        ]
+      ],
+      group: [Sequelize.col(`Category.id`)],
+      include: [{
+        model: this._articleCategoryModel,
+        as: Aliases.ARTICLES_CATEGORIES,
+        attributes: [],
+        required: true
+      }]
+    });
+
+    if (category) {
+      throw new Error(`Can not delete category with articles.`);
+    }
+
     const deletedRows = await this._categoryModel.destroy({where: {'id': categoryId}});
     return !!deletedRows;
   }
